+++ adapter/pyproject.toml
@@ -10,6 +10,7 @@ packages = [
 [tool.poetry.dependencies]
 python = ">=3.7.1, <3.11"
 dbt-core = "1.3.0b2"
+fal = ">=0.6.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
