+++ adapter/src/dbt/adapters/fal/connections.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 
 from dbt.adapters.base import Credentials
-from faldbt.adapters.python import PythonConnectionManager
+from dbt.adapters.fal.python import PythonConnectionManager
 
 
 class FalConnectionManager(PythonConnectionManager):
