+++ adapter/src/dbt/adapters/fal/python/__init__.py
@@ -1,0 +1,3 @@
+# these are all just exports, #noqa them so flake8 will be happy
+from dbt.adapters.fal.python.connections import PythonConnectionManager  # noqa
+from dbt.adapters.fal.python.impl import PythonAdapter  # noqa
