+++ adapter/src/dbt/adapters/fal/support/duckdb.py
