+++ adapter/src/dbt/adapters/fal/support/snowflake.py
