--- src/faldbt/adapters/python/__init__.py
@@ -1,3 +1,0 @@
-# these are all just exports, #noqa them so flake8 will be happy
-from faldbt.adapters.python.connections import PythonConnectionManager  # noqa
-from faldbt.adapters.python.impl import PythonAdapter  # noqa
